# Remove unsafe from Uri.HexEscape in favor of string.Create