@danmosemsft 2018-02-16
You have regex work in here