@ViktorHofer 2018-02-16
@dotnet-bot test NETFX x86 Release Build