@ViktorHofer 2018-02-16
@stephentoub I would expect the delegate to be allocated on the heap but the benchmark doesn't show that...