@stephentoub 2018-02-16
It doesn't close over any instance state, so the compiler caches it.
https://blogs.msdn.microsoft.com/pfxteam/2012/02/03/know-thine-implicit-allocations/